% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{cache_path}
\alias{cache_path}
\title{httr2 cache location}
\usage{
cache_path()
}
\description{
When opted-in to, httr2 caches OAuth tokens in this directory. By default,
it uses a OS-standard cache directory, but, if needed, you can override the
location by setting the \code{HTTR2_CACHE} env var.
}
\keyword{internal}
