% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{calculate_n_pages}
\alias{calculate_n_pages}
\title{Calculate the number of pages}
\usage{
calculate_n_pages(page_size, total)
}
\arguments{
\item{page_size}{An integer.}

\item{total}{A function that extracts the total number of elements from the \link{response}.}
}
\value{
A function that can be passed to the \code{n_pages} argument of \code{\link[=req_paginate]{req_paginate()}}.
}
\description{
Calculate the number of pages
}
\examples{
calculate_n_pages(page_size = 100, total = 250)
}
