% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_req_perform.R
\name{chunk_req_perform}
\alias{chunk_req_perform}
\title{Perform a request in chunks}
\usage{
chunk_req_perform(req, data, chunk_size, apply_chunk, parse_resp = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{data}{The data to chunk.}

\item{chunk_size}{The size of each chunk.}

\item{apply_chunk}{A function that applies the chunk to the request. It
takes two arguments:
\enumerate{
\item \code{req}: the original request.
\item \code{chunk}: the current data chunk.
}}

\item{parse_resp}{A function with one argument \code{resp} that parses the
response.}
}
\value{
A list of parsed responses.
}
\description{
Perform a request in chunks
}
\examples{
ids <- 1:20
responses <- request("https://api.restful-api.dev/objects") \%>\%
  chunk_req_perform(
    chunk_size = 10,
    data = ids,
    req_prep = function(req, chunk) {
      chunk <- set_names(chunk, "id")
      req_url_query(req, !!!chunk)
    },
    parse_resp = function(resp) {
      # resp
      parsed <- resp_body_json(resp)
      data.frame(
        id = sapply(parsed, function(x) x$id),
        name = sapply(parsed, function(x) x$name)
      )
    }
  )

responses
}
