% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-auth-code.R
\name{default_redirect_uri}
\alias{default_redirect_uri}
\title{OAuth default redirect uri}
\usage{
default_redirect_uri()
}
\description{
Returns \verb{http://localhost} but also respects the \code{HTTR2_OAUTH_REDIRECT_URL}
environment variable.
}
\keyword{internal}
