% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_app_logged_in}
\alias{handle_oauth_app_logged_in}
\title{Handle OAuth for Logged-in App Users}
\usage{
handle_oauth_app_logged_in(
  req,
  client_config,
  require_auth,
  cookie,
  key,
  httpHandler
)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}

\item{require_auth}{Logical, whether authentication is required.}

\item{cookie}{The name of the cookie where the app's token is stored.}

\item{key}{A secret key used to encrypt and decrypt tokens.}

\item{httpHandler}{A function to handle HTTP requests after authentication.}
}
\value{
An HTTP response object or \code{NULL} if authentication fails.
}
\description{
This function processes OAuth requests for logged-in users of the app. It
checks for an existing OAuth token, and if none is found and authentication
is required, it returns \code{NULL}. Otherwise, it invokes the provided HTTP
handler to continue processing the request.
}
