% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_app_login}
\alias{handle_oauth_app_login}
\title{Handle OAuth Login for App}
\usage{
handle_oauth_app_login(req, client_config, login_ui)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}

\item{login_ui}{A UI function or object that provides a custom login page.}
}
\value{
An HTTP response object or \code{NULL} if the request path is not the
root.
}
\description{
This function handles the OAuth login process for the app. It checks if the
request path is the root ("/"), and if so, either redirects to the primary
authentication provider or displays a custom login UI. The custom UI can
include a message and login options.
}
\keyword{internal}
