% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_app_logout}
\alias{handle_oauth_app_logout}
\title{Handle OAuth Logout for App}
\usage{
handle_oauth_app_logout(req, client_config, logout_path, cookie, logout_ui)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}

\item{logout_path}{The path that triggers the app's logout process.}

\item{cookie}{The name of the cookie where the app's token is stored.}

\item{logout_ui}{A UI function or object that provides a custom logout page.}
}
\value{
An HTTP response object or \code{NULL} if the request path does not match
the logout path.
}
\description{
This function handles the logout process for the app. It deletes the relevant
cookies from the user's browser and returns an HTTP response that either
displays a logout UI or redirects the user directly.
}
\keyword{internal}
