% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_client_callback}
\alias{handle_oauth_client_callback}
\title{Handle OAuth Client Callback}
\usage{
handle_oauth_client_callback(
  req,
  client_config,
  require_auth,
  cookie,
  key,
  token_validity
)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}

\item{require_auth}{Logical, whether authentication is required.}

\item{cookie}{The name of the cookie where the app's token is stored.}

\item{key}{A secret key used to encrypt and decrypt tokens.}

\item{token_validity}{The duration for which the token should be cached.}
}
\value{
An HTTP response object or \code{NULL} if the callback parameters are
invalid.
}
\description{
This function processes the OAuth callback after the user has authorized the
app. It exchanges the authorization code for an access token, handles PKCE
verification, and sets cookies for both the app and the client access token.
}
\keyword{internal}
