% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_client_login}
\alias{handle_oauth_client_login}
\title{Handle OAuth Client Login}
\usage{
handle_oauth_client_login(req, client_config)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}
}
\value{
An HTTP response object or \code{NULL} if no client is matched.
}
\description{
This function initiates the OAuth login process for clients configured with
the app. It loops through the client configurations and redirects the user to
the appropriate OAuth authorization URL.
}
\keyword{internal}
