% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_client_login_redirect}
\alias{handle_oauth_client_login_redirect}
\title{Handle OAuth Client Login Redirect}
\usage{
handle_oauth_client_login_redirect(req, client)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client}{A single client configuration object.}
}
\value{
An HTTP response object or \code{NULL} if the request path does not match
the client's login path.
}
\description{
This function handles redirection to the OAuth authorization URL for a
specific client. It manages PKCE (Proof Key for Code Exchange) if enabled,
and sets up necessary cookies before redirecting the user.
}
