% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{in_query}
\alias{in_query}
\alias{in_header}
\alias{in_body}
\title{Specify a request parameter}
\usage{
in_query(name, value = NULL, error_call = caller_env())

in_header(name, value = NULL, error_call = caller_env())

in_body(path, value = NULL, error_call = caller_env())
}
\arguments{
\item{name}{Name of the query resp. header parameter.}

\item{value}{The value of the parameter.}

\item{error_call}{}

\item{path}{Path of the parameter in the body.}
}
\value{
A parameter object.
}
\description{
Specify a request parameter
}
\examples{
in_query("start", 20)
}
