% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-online.R
\name{is_online}
\alias{is_online}
\title{Is your computer currently online?}
\usage{
is_online()
}
\description{
This function uses some cheap heuristics to determine if your computer is
currently online. It's a simple wrapper around \code{\link[curl:nslookup]{curl::has_internet()}}
exported from httr2 for convenience.
}
