% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{iterate_next_request}
\alias{iterate_next_request}
\title{Retrieve the next request from an iterative response}
\usage{
iterate_next_request(req, parsed)
}
\arguments{
\item{req}{A \link{request}.}

\item{parsed}{The response parsed by the argument \code{parse_resp} of \code{\link[=req_paginate]{req_paginate()}}.}
}
\value{
Generates the next request in an iterative request,
or \code{NULL} if there are no more pages to return.
}
\description{
In most case you should use \code{\link[=req_perform_iteratively]{req_perform_iteratively()}} but you can
use this lower-level helper if you need more control.
}
\examples{
req_flowers <- request(example_url()) \%>\%
  req_url_path("/iris") \%>\%
  req_url_query(limit = 40) \%>\%
  req_paginate_page_index(
    page_index = function(req, page) {
      req \%>\% req_url_query(page_index = page)
    }
  )
req_flowers \%>\% .$url

resp <- req_flowers \%>\% req_perform()
next_req <- iterate_next_request(req_flowers, resp)
next_req \%>\% .$url
}
\keyword{internal}
