% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-mock.R
\name{mock_resp_raw}
\alias{mock_resp_raw}
\alias{mock_resp_lines}
\title{mocking a connection object w/o digging into C code?}
\usage{
mock_resp_raw(resp, mocks)

mock_resp_lines(resp, mocks)
}
\arguments{
\item{resp}{A streaming \link{response} created by \code{\link[=req_perform_connection]{req_perform_connection()}}.}

\item{mocks}{A list with objects of class \code{raw}}
}
\value{
httr2 response
}
\description{
mocking a connection object w/o digging into C code?
}
\examples{
# resp_stream_raw
raw_mock <- function(req) {
  list(
    charToRaw("hello world"),
    charToRaw("hello mars"),
    charToRaw("hello jupiter")
  )
}

# with context manager
with_mocked_responses(raw_mock, {
  resp <- request(example_url()) |>
    req_url_path("/stream-bytes/32768") |>
    req_perform_connection()
  print(resp_stream_raw(resp))
  print(resp_stream_raw(resp))
  print(resp_stream_raw(resp))
  print(resp_stream_raw(resp))
  close(resp)
})

# local version
local_mocked_responses(raw_mock)
resp <- request(example_url()) |>
  req_url_path("/stream-bytes/32768") |>
  req_perform_connection()
resp_stream_raw(resp)
resp_stream_raw(resp)
resp_stream_raw(resp)
resp_stream_raw(resp)
close(resp)

# resp_stream_lines
lines_mock <- function(req) {
  list(
    '{"a": 1}',
    '{"b": 2}',
    '{"c": 3}'
  )
}

# with context manager
with_mocked_responses(lines_mock, {
  resp <- request(example_url()) |>
    req_template("GET /stream/:n", n = 3) |>
    req_perform_connection()
  print(resp_stream_lines(resp))
  print(resp_stream_lines(resp))
  print(resp_stream_lines(resp))
  print(resp_stream_lines(resp))
  close(resp)
})

# local version
local_mocked_responses(lines_mock)
resp <- request(example_url()) |>
  req_template("GET /stream/:n", n = 3) |>
  req_perform_connection()
resp_stream_lines(resp)
resp_stream_lines(resp)
resp_stream_lines(resp)
resp_stream_lines(resp)
close(resp)
}
\keyword{internal}
