% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-app.R
\name{oauth_app}
\alias{oauth_app}
\alias{oauth_client}
\title{Create an OAuth app}
\usage{
oauth_app(client, endpoints, auth = c("body", "header"))

oauth_client(client_id, client_secret = NULL, name = NULL)
}
\arguments{
\item{client}{A OAuth client created with \code{oauth_client()}.}

\item{endpoints}{A named character vector of endpoints. The precise endpoints
required depend on the flow that you'll use, but all require a \code{token}
endpoint that returns an access token.}

\item{auth}{Authentication mechanism used by the API to authenticate
confidential requests made during the flow. This is most commonly
\code{"body"} where the \code{client_id} and \code{client_secret} are added to the body
of the request, but can also be \code{"header"} where the \code{client_id}
and \code{client_secret} are sent using the HTTP Authorization header.}

\item{client_id}{Client identifier.}

\item{client_secret}{Client secret. This is technically confidential so you
should avoid storing it in source code where possible. However, many APIs
require it in order to provide a user friendly authentication experience,
and the risks of including it are usually low. To make things a little
safer, I recommend using \code{\link[=obfuscate]{obfuscate()}} to store an obfuscated version of
the source.}

\item{name}{Optional name for the client. Used when generating cache
directory. If \code{NULL}, generated from hash of app hostname and
\code{client_id}.}
}
\description{
An OAuth app is the combination of a client, a set of endpoints
(i.e. urls where various requests should be sent), and an authentication
mechanism. A client consists of at least a \code{client_id}, and also often
a \code{client_secret}. You'll get these values when you create the client on
the API's website.
}
