% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_cache_clear}
\alias{oauth_cache_clear}
\title{Clear OAuth cache}
\usage{
oauth_cache_clear(client, cache_disk = FALSE, cache_key = NULL)
}
\arguments{
\item{client}{An \code{\link[=oauth_client]{oauth_client()}}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

Learn more in \url{https://httr2.r-lib.org/articles/oauth.html}.}

\item{cache_key}{If you want to cache multiple tokens per app, use this
key to disambiguate them.}
}
\description{
Use this function to clear cached credentials.
}
