% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_cache_path}
\alias{oauth_cache_path}
\title{httr2 OAuth cache location}
\usage{
oauth_cache_path()
}
\description{
When opted-in to, httr2 caches OAuth tokens in this directory. By default,
it uses a OS-standard cache directory, but, if needed, you can override the
location by setting the \code{HTTR2_OAUTH_CACHE} env var.
}
