% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth_cache_prune}
\alias{oauth_cache_prune}
\title{Prune the OAuth token cache}
\usage{
oauth_cache_prune(max_age_days = 30)
}
\arguments{
\item{max_age_days}{Delete cached tokens that haven't been modified in
this many days.}
}
\description{
Deletes cached OAuth tokens (from both the current and legacy cache
directories, see \code{\link[=oauth_cache_path]{oauth_cache_path()}}) that are older than
\code{max_age_days}. This is called automatically when httr2 is loaded, so
you should only need to call it yourself if you want to prune the cache
immediately.
}
