% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-jwt.R
\name{oauth_flow_jwt}
\alias{oauth_flow_jwt}
\title{OAuth flow: JWT}
\usage{
oauth_flow_jwt(
  app,
  claims,
  signature,
  signature_params = list(),
  scope = NULL,
  token_params = list()
)
}
\arguments{
\item{app}{An \code{\link[=oauth_app]{oauth_app()}}.}

\item{claims}{A list of claims. If all elements of the claim set are static
apart from \code{iat}, \code{nbf}, \code{exp}, or \code{jti}, provide a list and
\code{\link[=jwt_claim_set]{jwt_claim_set()}} will automatically fill in the dynamic components.
If other components need to vary, you can instead provide a zero-argument
callback function which should call \code{jwt_claim_set()}.}

\item{signature}{Function use to sign \code{claim_set}, e.g. \code{\link[=jwt_sign_rs256]{jwt_sign_rs256()}}.}

\item{signature_params}{Additional arguments passed to \code{signature}}

\item{scope}{Scopes to be requested from the resource owner.}

\item{token_params}{List containing additional parameters passed to \code{oauth_flow_access_token()}.}
}
\description{
This function implements the OAuth client credentials flow, as defined
by \href{https://datatracker.ietf.org/doc/html/rfc7523#section-2.1}{rfc7523},
Section 2.1. It is often used for service accounts, accounts that are
used primarily in automated environments.
}
\seealso{
Other OAuth flows: 
\code{\link{oauth_flow_auth_code}()},
\code{\link{oauth_flow_client_credentials}()},
\code{\link{oauth_flow_device}()},
\code{\link{oauth_flow_password}()},
\code{\link{oauth_flow_refresh}()}
}
\concept{OAuth flows}
