% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-refresh.R
\name{oauth_flow_refresh}
\alias{oauth_flow_refresh}
\title{OAuth flow: refresh token}
\usage{
oauth_flow_refresh(
  app,
  refresh_token = Sys.getenv("HTTR_REFRESH_TOKEN"),
  scope = NULL,
  token_params = list()
)
}
\arguments{
\item{app}{An \code{\link[=oauth_app]{oauth_app()}}.}

\item{refresh_token}{A refresh token. This is equivalent to a password
so shouldn't be typed into the console or stored in a script. Instead,
we recommend placing in an environment variable; the default behaviour
is to look in \code{HTTR_REFRESH_TOKEN}.}

\item{scope}{Scopes to be requested from the resource owner.}

\item{token_params}{List containing additional parameters passed to \code{oauth_flow_access_token()}.}
}
\description{
This function generates an access token from a refresh token, following
the process described in
\href{https://datatracker.ietf.org/doc/html/rfc6749#section-6}{rfc6749},
Section 6. This is primarily useful for testing - you can manually performed
some other OAuth flow (e.g. \code{\link[=oauth_flow_auth_code]{oauth_flow_auth_code()}} or
\code{\link[=oauth_flow_device]{oauth_flow_device()}}) and save the refresh token into an environment
variable.

When requesting an access token, the server may also return a new refresh
token. If this happens, \code{oauth_flow_refresh()} will error, and you'll have
to create a new refresh token following the same procedure you did to get
the first token (so it's a good idea to document what you did the first time
because you might need to do it again).
}
\seealso{
Other OAuth flows: 
\code{\link{oauth_flow_auth_code}()},
\code{\link{oauth_flow_client_credentials}()},
\code{\link{oauth_flow_device}()},
\code{\link{oauth_flow_password}()}
}
\concept{OAuth flows}
