% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-auth-code.R
\name{oauth_redirect_uri}
\alias{oauth_redirect_uri}
\title{Default redirect url for OAuth}
\usage{
oauth_redirect_uri()
}
\description{
The default redirect uri used by \code{\link[=req_oauth_auth_code]{req_oauth_auth_code()}}. Defaults to
\verb{http://localhost} unless the \code{HTTR2_OAUTH_REDIRECT_URL} envvar is set.
}
