% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-example.R
\name{oauth_shiny_app_example}
\alias{oauth_shiny_app_example}
\title{Example Shiny App Using OAuth}
\usage{
oauth_shiny_app_example(
  client_config,
  require_auth = TRUE,
  key = oauth_shiny_app_passphrase(),
  dark_mode = FALSE,
  login_ui = oauth_shiny_ui_login(client_config, dark_mode),
  logout_ui = oauth_shiny_ui_logout(client_config, dark_mode),
  logout_path = "logout",
  logout_on_token_expiry = FALSE,
  cookie_name = "oauth_app_token",
  token_validity = 3600,
  redact = TRUE
)
}
\arguments{
\item{client_config}{A \code{oauth_shiny_config} object containing a list of OAuth
client configurations (\code{oauth_shiny_client} objects).}

\item{require_auth}{Logical; should login be enforced? Defaults to \code{FALSE}.}

\item{key}{The encryption key for securing cookies. Defaults to the
environment variable \code{HTTR2_OAUTH_PASSPHRASE}. It should be a long,
randomly generated string, which can be created using
\code{httr2::secret_make_key()} or a similar method.}

\item{dark_mode}{Logical; should the login and logout UI use dark mode?
Defaults to \code{FALSE}..}

\item{login_ui}{The login UI to present to users when \code{enforce_login = TRUE}.
Defaults to an automatically generated UI based on the clients listed in
\code{client_config}.}

\item{logout_ui}{The logout UI to present to users. Defaults to an
automatically generated UI based on the providers listed in
\code{client_config}.}

\item{logout_path}{The URL path used to log users out of the app. Defaults to
\code{'logout'}.}

\item{logout_on_token_expiry}{Logical; should the user be automatically
logged out of the app when the token expires? If \code{FALSE}, the session
remains active until the user refreshes the browser or manually logs out.
Defaults to \code{FALSE}.}

\item{cookie_name}{The name of the cookie used for authentication. Defaults
to \code{'oauth_app_token'}.}

\item{token_validity}{Validity of the app token in seconds. Defaults to
1 hour (3600s) after which the token and cookie expires.}

\item{redact}{Logical, whether to redact tokens in UI. Defaults to \code{TRUE}.}
}
\value{
A configured Shiny app object that allows users to authenticate using
the specified OAuth providers.
}
\description{
This function sets up a Shiny application that uses multiple OAuth providers
for authentication. The application can be configured to allow users to sign
in using various OAuth providers like Google, Microsoft, GitHub, and others.
}
\examples{
\dontrun{
config <- oauth_shiny_client_config(
  oauth_shiny_client_github(),
  oauth_shiny_client_google()
)

oauth_shiny_app_example(config, dark_mode = TRUE)
}
}
