% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-app.R
\name{oauth_shiny_app_url}
\alias{oauth_shiny_app_url}
\title{Override app url for OAuth}
\usage{
oauth_shiny_app_url()
}
\description{
It can be difficult to correctly infer the correct app url depending on
which environment the app is running in (localhost, shinyapps, cloud, etc).
httr2 makes an attempt to guess the correct app url, but the environment
variable \code{HTTR2_OAUTH_APP_URL} could be used to override a wrong guess.
}
