% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{oauth_shiny_callback_resolve_client}
\alias{oauth_shiny_callback_resolve_client}
\title{Resolve OAuth Client in Callback}
\usage{
oauth_shiny_callback_resolve_client(client_config, query)
}
\arguments{
\item{client_config}{A list of client configurations used for OAuth.}

\item{query}{A list of query parameters from the OAuth callback request.}
}
\value{
The matching client configuration object.
}
\description{
This function resolves which OAuth client configuration should be used during
the callback phase based on the query parameters returned by the OAuth
server. It ensures that the client matches either the state parameter or the
OpenID issuer URL.
}
\keyword{internal}
