% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-client.R
\name{oauth_shiny_client_config}
\alias{oauth_shiny_client_config}
\title{Configure Multiple OAuth Shiny Clients}
\usage{
oauth_shiny_client_config(...)
}
\arguments{
\item{...}{A series of OAuth Shiny Client objects created by
\code{oauth_shiny_client()} or its specialized versions.}
}
\value{
A named list of OAuth Shiny Client objects. The names of the list
elements correspond to the \code{name} parameter of each client.
}
\description{
This function allows you to configure multiple OAuth Shiny Clients by
combining them into a single list. This is useful when your Shiny application
supports multiple OAuth providers (e.g., Google, Microsoft, GitHub, etc.).
}
\details{
Each client in the list must be an object created by the
\code{oauth_shiny_client()} function or one of its specialized versions (e.g.,
\code{oauth_shiny_client_google()}).
}
\examples{
\dontrun{
google_client <- oauth_shiny_client_google()
microsoft_client <- oauth_shiny_client_microsoft()
config <- oauth_shiny_client_config(google_client, microsoft_client)
}

}
