% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-client.R
\name{oauth_shiny_client_github}
\alias{oauth_shiny_client_github}
\title{Create a Github OAuth Client in Shiny}
\usage{
oauth_shiny_client_github(
  name = "github",
  id = Sys.getenv("OAUTH_GITHUB_CLIENT_ID"),
  secret = Sys.getenv("OAUTH_GITHUB_CLIENT_SECRET"),
  redirect_uri = oauth_redirect_uri(),
  scope = "read:user user:email",
  login_button = oauth_shiny_ui_button_github(),
  login_button_dark = oauth_shiny_ui_button_github_dark(),
  auth_set_custom_claim = oauth_shiny_client_github_set_custom_claim,
  ...
)
}
\arguments{
\item{name}{The name of the OAuth Provider, default is "github".}

\item{id}{Client ID, defaults to environment variable
\code{OAUTH_GITHUB_CLIENT_ID}.}

\item{secret}{Client Secret, defaults to environment variable
\code{OAUTH_GITHUB_CLIENT_SECRET}.}

\item{redirect_uri}{Authorization callback URL, defaults to
\code{oauth_redirect_uri()}.}

\item{scope}{Scopes to be requested, defaults to "read:user user:email".}

\item{login_button}{Button used for GitHub login, defaults to
\code{oauth_shiny_ui_button_github()}.}

\item{login_button_dark}{Dark theme button for GitHub login, defaults to
\code{oauth_shiny_ui_button_github_dark()}.}

\item{auth_set_custom_claim}{Custom function for setting claims, defaults
to \code{oauth_shiny_client_github_set_custom_claim}.}

\item{...}{Additional arguments passed to \code{oauth_shiny_client()}.}
}
\value{
A Shiny OAuth Client configured for GitHub.
}
\description{
This function creates an OAuth Shiny Client specifically for GitHub. It uses
the general \code{oauth_shiny_client} function to set up the client with
GitHub-specific defaults. Note that Github does not currently support openid,
hence it requires additional requests to \verb{/user/email} to authorize the user.
Create the app from your \href{https://github.com/settings/developers}{Developer Settings}
}
\examples{
\dontrun{
spotify_client <- oauth_shiny_client_spotify()
}

}
