% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-client.R
\name{oauth_shiny_client_google}
\alias{oauth_shiny_client_google}
\title{Create an OAuth Shiny Client for Google}
\usage{
oauth_shiny_client_google(
  name = "google",
  id = Sys.getenv("OAUTH_GOOGLE_CLIENT_ID"),
  secret = Sys.getenv("OAUTH_GOOGLE_CLIENT_SECRET"),
  openid_issuer_url = "https://accounts.google.com/",
  scope = "openid profile email",
  login_button = oauth_shiny_ui_button_google(),
  login_button_dark = oauth_shiny_ui_button_google_dark(),
  ...
)
}
\arguments{
\item{name}{The name of the OAuth Provider, default is "google".}

\item{id}{Client ID, defaults to environment variable
\code{OAUTH_GOOGLE_CLIENT_ID}.}

\item{secret}{Client Secret, defaults to environment variable
\code{OAUTH_GOOGLE_CLIENT_SECRET}.}

\item{openid_issuer_url}{URL for OpenID issuer, defaults to
"https://accounts.google.com/".}

\item{scope}{Scopes to be requested, defaults to "openid profile email".}

\item{login_button}{Button used for Google login, defaults to
\code{oauth_shiny_ui_button_google()}.}

\item{login_button_dark}{Dark theme button for Google login, defaults to
\code{oauth_shiny_ui_button_google_dark()}.}

\item{...}{Additional arguments passed to \code{oauth_shiny_client()}.}
}
\value{
A Shiny OAuth Client configured for Google.
}
\description{
This function creates an OAuth Shiny Client specifically for Google. It uses
the general \code{oauth_shiny_client} function to set up the client with
Google-specific defaults.
}
\details{
For parameter details, see \code{\link[=oauth_shiny_client]{oauth_shiny_client()}}.
}
\examples{
\dontrun{
google_client <- oauth_shiny_client_google()
}

}
