% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-client.R
\name{oauth_shiny_client_microsoft}
\alias{oauth_shiny_client_microsoft}
\title{Create an OAuth Shiny Client for Microsoft}
\usage{
oauth_shiny_client_microsoft(
  name = "microsoft",
  id = Sys.getenv("OAUTH_MICROSOFT_CLIENT_ID"),
  secret = Sys.getenv("OAUTH_MICROSOFT_CLIENT_SECRET"),
  openid_issuer_url,
  scope = "openid profile email",
  login_button = oauth_shiny_ui_button_microsoft(),
  login_button_dark = oauth_shiny_ui_button_microsoft_dark(),
  ...
)
}
\arguments{
\item{name}{The name of the OAuth Provider, default is "microsoft".}

\item{id}{Client ID, defaults to environment variable
\code{OAUTH_MICROSOFT_CLIENT_ID}.}

\item{secret}{Client Secret, defaults to environment variable
\code{OAUTH_MICROSOFT_CLIENT_SECRET}.}

\item{openid_issuer_url}{URL for OpenID issuer.}

\item{scope}{Scopes to be requested, defaults to "openid profile email".}

\item{login_button}{Button used for Microsoft login, defaults to
\code{oauth_shiny_ui_button_microsoft()}.}

\item{login_button_dark}{Dark theme button for Microsoft login, defaults to
\code{oauth_shiny_ui_button_microsoft_dark()}.}

\item{...}{Additional arguments passed to \code{oauth_shiny_client()}.}
}
\value{
A Shiny OAuth Client configured for Microsoft.
}
\description{
This function creates an OAuth Shiny Client specifically for Microsoft. It
uses the general \code{oauth_shiny_client} function to set up the client with
Microsoft-specific defaults.
}
\details{
For parameter details, see \code{\link[=oauth_shiny_client]{oauth_shiny_client()}}.
}
\examples{
\dontrun{
microsoft_client <- oauth_shiny_client_microsoft()
}

}
