% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-client.R
\name{oauth_shiny_client_spotify}
\alias{oauth_shiny_client_spotify}
\title{This function creates an OAuth Shiny Client specifically for Spotify. It uses
the general \code{oauth_shiny_client} function to set up the client with
Spotify-specific defaults.}
\usage{
oauth_shiny_client_spotify(
  name = "spotify",
  id = Sys.getenv("OAUTH_SPOTIFY_CLIENT_ID"),
  secret = Sys.getenv("OAUTH_SPOTIFY_CLIENT_SECRET"),
  redirect_uri = oauth_redirect_uri(),
  scope = "user-read-email",
  login_button = oauth_shiny_ui_button_spotify(),
  login_button_dark = oauth_shiny_ui_button_spotify_dark(),
  auth_set_custom_claim = oauth_shiny_client_spotify_set_custom_claim,
  ...
)
}
\arguments{
\item{name}{The name of the OAuth Provider, default is "spotify".}

\item{id}{Client ID, defaults to environment variable
\code{OAUTH_SPOTIFY_CLIENT_ID}.}

\item{secret}{Client Secret, defaults to environment variable
\code{OAUTH_SPOTIFY_CLIENT_SECRET}.}

\item{redirect_uri}{Authorization callback URL, defaults to
\code{oauth_redirect_uri()}.}

\item{scope}{Scopes to be requested, defaults to "user-read-email".}

\item{login_button}{Button used for Spotify login, defaults to
\code{oauth_shiny_ui_button_spotify()}.}

\item{login_button_dark}{Dark theme button for Spotify login, defaults to
\code{oauth_shiny_ui_button_spotify_dark()}.}

\item{auth_set_custom_claim}{Custom function for setting claims, defaults
to \code{oauth_shiny_client_spotify_set_custom_claim}.}

\item{...}{Additional arguments passed to \code{oauth_shiny_client()}.}
}
\value{
A Shiny OAuth Client configured for Spotify.
}
\description{
For parameter details, see \code{\link[=oauth_shiny_client]{oauth_shiny_client()}}.
}
\examples{
\dontrun{
spotify_client <- oauth_shiny_client_spotify()
}
}
