% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_decrypt_client_token}
\alias{oauth_shiny_decrypt_client_token}
\title{Decrypt OAuth Access Token}
\usage{
oauth_shiny_decrypt_client_token(cookie, key)
}
\arguments{
\item{cookie}{A string representing the encrypted cookie value.}

\item{key}{A secret key used to decrypt the cookie.}
}
\value{
A decrypted OAuth token object.
}
\description{
This function decrypts an OAuth access token stored in a cookie.
}
