% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_encrypt_client_token}
\alias{oauth_shiny_encrypt_client_token}
\title{Encrypt OAuth Client Access Token}
\usage{
oauth_shiny_encrypt_client_token(client, token, key, max_age)
}
\arguments{
\item{client}{A list representing the client configuration, including the
client secret.}

\item{token}{A list representing the OAuth token to be encrypted.}

\item{key}{A secret key used to encrypt the token.}

\item{max_age}{Integer specifying the expiration time of the token in
seconds.}
}
\value{
A string representing the encrypted token in hexadecimal format.
}
\description{
This function encrypts an OAuth access token for storage in a cookie.
}
