% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_get_access_token}
\alias{oauth_shiny_get_access_token}
\title{Get OAuth Client Access Token from Shiny App Cookie}
\usage{
oauth_shiny_get_access_token(
  client,
  key = oauth_shiny_app_passphrase(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{client}{A list representing the client configuration,
including the client name and secret.}

\item{key}{A secret key used to decrypt the token.}

\item{session}{The current Shiny session object, used to access request
details. Default is \code{shiny::getDefaultReactiveDomain()}.}
}
\value{
A decrypted \code{oauth_token} if the cookie exists and is valid
}
\description{
This function retrieves and decrypts an OAuth token stored in a cookie.
It automatically handles chunked cookies if the encrypted token exceeds
maximum cookie size (e.g. \code{oauth_app_google_token_1} and
\code{oauth_app_google_token2})
}
