% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_get_app_token}
\alias{oauth_shiny_get_app_token}
\title{Get OAuth Token from Shiny App Cookie}
\usage{
oauth_shiny_get_app_token(
  cookie = "oauth_app_token",
  key = oauth_shiny_app_passphrase(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{cookie}{A string representing the name of the cookie where the JWT is
stored. Default is \code{oauth_app_token}.}

\item{key}{The secret key used in the application.}

\item{session}{Defaults to \code{shiny::getDefaultReactiveDomain()}}
}
\value{
A decoded JWT if the cookie exists and is valid; otherwise, \code{NULL}.
}
\description{
This function retrieves and decodes the JWT stored in a specified cookie from
the incoming request.
}
