% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-app.R
\name{oauth_shiny_infer_app_url}
\alias{oauth_shiny_infer_app_url}
\title{Extract server URL from the request}
\usage{
oauth_shiny_infer_app_url(req)
}
\arguments{
\item{req}{A request object.}
}
\value{
The app url.
}
\description{
Inferring the correct app url on the server requires some work.
This function attempts to guess the correct server url, but may fail outside
of tested hosts (\verb{127.0.0.1} and \code{shinyapps.io}). To be sure, set the
environment variable \code{HTTR2_OAUTH_APP_URL} explicitly. Logic inspired by
\href{r4ds/shinyslack}{https://github.com/r4ds/shinyslack}.
}
\keyword{internal}
