% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_set_access_token}
\alias{oauth_shiny_set_access_token}
\title{Set OAuth Client Access Token in Shiny App Cookie}
\usage{
oauth_shiny_set_access_token(client, token, key)
}
\arguments{
\item{client}{An \code{oauth_shiny_client} representing the client configuration}

\item{token}{A list representing the OAuth token to be stored.}

\item{key}{A secret key used for encrypting the token.}
}
\value{
None. The function sets the encrypted token in the HTTP response
header as a cookie.
}
\description{
This function stores an OAuth access token in a cookie, encrypting it before
storage. The token is an \code{oauth_token} object containing access_token and
related information.
}
