% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-token.R
\name{oauth_shiny_set_app_token}
\alias{oauth_shiny_set_app_token}
\title{Set OAuth Token in Shiny App Cookie}
\usage{
oauth_shiny_set_app_token(claims, cookie, key, token_validity)
}
\arguments{
\item{claims}{A list containing the claims to be included in the JWT.}

\item{cookie}{A string representing the name of the cookie where the JWT will
be stored.}

\item{key}{A secret key used to sign the JWT.}

\item{token_validity}{Integer specifying the expiration time of the JWT in
seconds.}
}
\value{
None. The function sets the cookie in the HTTP response header.
}
\description{
This function sets a JSON Web Token (JWT) and stores it in a cookie so the
user can be successfully identified and skip login when returning to the app.
The token is stored in a httponly, secure cookie and signed with HMAC. Claims
include \code{identifier}, \code{name}, \code{email}, \code{aud} and \code{sub}.
}
