% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-ui.R
\name{oauth_shiny_ui_button}
\alias{oauth_shiny_ui_button}
\title{Create a Custom OAuth Shiny UI Login Button}
\usage{
oauth_shiny_ui_button(path, title, logo = NULL, theme = c("light", "dark"))
}
\arguments{
\item{path}{The URL path for the OAuth login.}

\item{title}{The title to display on the button.}

\item{logo}{The file path or SVG code for the logo to display on the button.}

\item{theme}{The theme of the button, either "light" or "dark". Defaults to
"light".}
}
\value{
A Shiny UI object representing the login button.
}
\description{
This function creates a custom login button for an OAuth provider, allowing
customization of the button's appearance.
}
