% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-ui.R
\name{oauth_shiny_ui_login}
\alias{oauth_shiny_ui_login}
\title{Create an OAuth Shiny UI Login Page}
\usage{
oauth_shiny_ui_login(oauth_shiny_client_config, dark_mode = FALSE)
}
\arguments{
\item{oauth_shiny_client_config}{A list of client configurations for OAuth
providers.}

\item{dark_mode}{A logical value indicating whether to use dark mode.
Defaults to \code{FALSE}.}
}
\value{
A Shiny UI object representing the login page.
}
\description{
This function creates the Shiny UI for the OAuth login page, allowing users
to log in with their preferred OAuth provider. The UI can be customized for
dark mode.
}
