% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-ui.R
\name{oauth_shiny_ui_logout}
\alias{oauth_shiny_ui_logout}
\title{Create an OAuth Shiny UI Logout Page}
\usage{
oauth_shiny_ui_logout(oauth_shiny_client_config, dark_mode = FALSE)
}
\arguments{
\item{oauth_shiny_client_config}{A list of client configurations for OAuth
providers.}

\item{dark_mode}{A logical value indicating whether to use dark mode.
Defaults to \code{FALSE}.}
}
\value{
A Shiny UI object representing the logout page.
}
\description{
This function creates the Shiny UI for the OAuth logout page, displaying a
message that the user has successfully logged out.
}
