% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate_perform}
\alias{paginate_perform}
\alias{paginate_next_request}
\title{Perform a paginated request}
\usage{
paginate_perform(req, max_pages = 20L, progress = TRUE)

paginate_next_request(resp, req)
}
\arguments{
\item{req}{A \link{request}.}

\item{max_pages}{The maximum number of pages to request.}

\item{progress}{Display a progress bar?}

\item{resp}{An HTTP \link{response}.}
}
\value{
A list of responses.
}
\description{
Perform a paginated request
}
\examples{
req_pokemon <- request("https://pokeapi.co/api/v2/pokemon") \%>\%
  req_url_query(limit = 150) \%>\%
  req_paginate_next_url(
    next_url = function(resp) resp_body_json(resp)[["next"]],
    n_pages = function(resp) {
      calculate_n_pages(
        page_size = 150,
        total = resp_body_json(resp)$count
      )
    }
  )

responses <- paginate_perform(req_pokemon)
}
