% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{req}
\alias{req}
\title{Create a new HTTP request}
\usage{
req(base_url)
}
\arguments{
\item{base_url}{Base URL for request.}
}
\description{
To perform a HTTP request, first create a request object with \code{req()},
then define its behaviour with \code{req_} functions, then perform the request
and fetch the response with \code{\link[=req_fetch]{req_fetch()}}.
}
\examples{
req("http://r-project.org")
}
