% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-auth.R
\name{req_auth_basic}
\alias{req_auth_basic}
\title{Authenticate request with HTTP basic authentication}
\usage{
req_auth_basic(req, username, password = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{username}{User name.}

\item{password}{Password. You avoid entering the password directly when
calling this function as it will be captured by \code{.Rhistory}. Instead,
leave it unset and the default behaviour will prompt you for it
interactively.}
}
\description{
This sets the Authorization header. See details at
\url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Authorization}.
}
\examples{
request("http://example.com") \%>\%
  req_auth_basic("hadley", "my-secret-password") \%>\%
  req_dry_run()

# Note that the authorization header is not encrypted and the
# password can easily be restored:
rawToChar(jsonlite::base64_dec("aGFkbGV5Om15LXNlY3JldC1wYXNzd29yZA=="))
# This means that you should be careful when sharing any code that
# reveals the Authorization header
}
