% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-options.R
\name{req_authenticate}
\alias{req_authenticate}
\title{Use basic HTTP authentication}
\usage{
req_authenticate(req, user_name, password, type = "basic")
}
\arguments{
\item{req}{A \link{req}uest.}

\item{user_name, password}{User name and password pair.}

\item{type}{Type of HTTP authentication. Should be one of the following
values: "basic", "digest", "digest_ie", "gssnegotiate", "ntlm", "any".}
}
\description{
This sets the Authorization header. See details at
\url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Authorization}.
}
\examples{
req("http://example.com") \%>\%
  req_authenticate("hadley", "my-secret-password") \%>\%
  req_dry_run()

# Note that the authorization header is not encrypted and the
# password can easily be restored:
rawToChar(jsonlite::base64_dec("aGFkbGV5Om15LXNlY3JldC1wYXNzd29yZA=="))
# This means that you should be careful when sharing any code that
# reveals the Authorization header
}
