% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-cookies.R
\name{req_cookie_file}
\alias{req_cookie_file}
\title{Preserve cookies across requests}
\usage{
req_cookie_file(req, path)
}
\arguments{
\item{req}{A \link{request}.}

\item{path}{A path to file where cookies will be saved.}
}
\description{
By default, httr2 uses a clean slate for every request meaning that cookies
are not automatically preserved across requests. To preserve cookies, you
must set a cookie file which will be read before and updated after each
request.
}
\examples{
path <- tempfile()
httpbin <- request(example_url()) \%>\%
  req_cookie_file(path)

# Manually set two cookies
httpbin \%>\%
  req_template("/cookies/set/:name/:value", name = "chocolate", value = "chip") \%>\%
  req_perform() \%>\%
  resp_body_json()

httpbin \%>\%
  req_template("/cookies/set/:name/:value", name = "oatmeal", value = "raisin") \%>\%
  req_perform() \%>\%
  resp_body_json()

# The cookie path has a straightforward format
readLines(path)
}
