% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-body.R
\name{req_get_body}
\alias{req_get_body}
\title{Retrieve the body that a request will send}
\usage{
req_get_body(req)
}
\arguments{
\item{req}{A httr2 \link{request} object.}
}
\description{
Mimics the algorithm that curl uses to determine the body.
}
\examples{
req <- request(example_url())
req |> req_body_raw("abc") |> req_get_body()
req |> req_body_file(system.file("DESCRIPTION")) |> req_get_body()
req |> req_body_json(list(x = 1, y = 2)) |> req_get_body()
req |> req_body_form(x = 1, y = 2) |> req_get_body()
req |> req_body_multipart(x = "x", y = "y") |> req_get_body() |> cat()
}
\keyword{internal}
