% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-url.R
\name{req_get_url}
\alias{req_get_url}
\title{Get request URL}
\usage{
req_get_url(req)
}
\arguments{
\item{req}{A httr2 \link{request} object.}
}
\value{
A character string.
}
\description{
Retrieve the URL from a request.
}
\examples{
request("https://httpbin.org") |>
  req_url_path("/get") |>
  req_url_query(hello = "world") |>
  req_get_url()
}
