% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-perform-multi.R
\name{req_next_multi}
\alias{req_next_multi}
\title{Create the next request of a multi request}
\usage{
req_next_multi(req, parsed)
}
\arguments{
\item{req}{A \link{request}.}

\item{parsed}{A parsed response.}
}
\value{
The next request.
}
\description{
Use this function to create the next request for a request that has a
multi request policy, e.g. created by \code{\link[=req_paginate]{req_paginate()}} or \code{\link[=req_chunk]{req_chunk()}}.
}
\examples{
ids <- 1:7
chunk_size <- 3

apply_chunk <- function(req, chunk) {
  chunk <- rlang::set_names(chunk, "id")
  req_url_query(req, !!!chunk)
}

req <- req_chunk(
  request("https://example.com"),
  ids,
  chunk_size = 3,
  apply_chunk = apply_chunk
)

req1 <- req_next_multi(req, parsed = NULL)
req1$url
req2 <- req_next_multi(req1, parsed = NULL)
req2$url
}
