% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-device.R
\name{req_oauth_device}
\alias{req_oauth_device}
\alias{oauth_flow_device}
\title{OAuth with device flow}
\usage{
req_oauth_device(
  req,
  client,
  auth_url,
  scope = NULL,
  auth_params = list(),
  token_params = list(),
  cache_disk = FALSE,
  cache_key = NULL
)

oauth_flow_device(
  client,
  auth_url,
  pkce = FALSE,
  scope = NULL,
  auth_params = list(),
  token_params = list()
)
}
\arguments{
\item{req}{A \link{request}.}

\item{client}{An \code{\link[=oauth_client]{oauth_client()}}.}

\item{auth_url}{Authorization url; you'll need to discover this by reading
the documentation.}

\item{scope}{Scopes to be requested from the resource owner.}

\item{auth_params}{A list containing additional parameters passed to
\code{\link[=oauth_flow_auth_code_url]{oauth_flow_auth_code_url()}}.}

\item{token_params}{List containing additional parameters passed to the
\code{token_url}.}

\item{cache_disk}{Should the access token be cached on disk? This reduces
the number of times that you need to re-authenticate at the cost of
storing access credentials on disk.

See the "Caching" section below for more details.}

\item{cache_key}{If you want to cache multiple tokens per app, use this
key to disambiguate them.}

\item{pkce}{Use "Proof Key for Code Exchange"? This adds an extra layer of
security and should always be used if supported by the server.}
}
\value{
\code{req_oauth_device()} returns a modified HTTP \link{request} that will
use OAuth; \code{oauth_flow_device()} returns an \link{oauth_token}.
}
\description{
Authenticate using the OAuth \strong{device flow}, as defined by \href{https://datatracker.ietf.org/doc/html/rfc8626}{RFC 8626}.
It's designed for devices that don't have access to a web browser (if you've
ever authenticated an app on your TV, this is probably the flow you've used),
but it also works well from within R.

Learn more about the overall OAuth authentication flow in \code{vignette("oauth")}.
}
\section{Caching}{
By default, the OAuth token will be cached in memory. That means that you
will only need to authenticate once in the current session, but you will
need to re-authenticate again from scratch if you restart R. In some cases,
you may want to save the token so that it's automatically used across
sessions. This is easy to do (just set \code{cache_disk = TRUE}) but you need to
carefully consider the consequences of saving the user's credentials on disk.

httr2 does the best it can to save these credentials securely. They are
stored in a local cache directory (\code{\link[=oauth_cache_path]{oauth_cache_path()}}) that should only
be accessible to the current user, and are encrypted so they will be hard
for any package other than httr2 to read. However, there's no way to prevent
other R code from using httr2 to access them, so if you do choose to cache
tokens, you should inform the user and give them the ability to opt-out.

You can see which clients have cached tokens by looking in the cache
directory used by httr2:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{dir(oauth_cache_path(), recursive = TRUE)
}\if{html}{\out{</div>}}

httr2 automatically deletes any cached tokens that are older than 30 days
whenever it's loaded. This means that you'll need to re-auth at least once a
month, but prevents tokens for hanging around on disk long after you've
forgotten you created them.
}

\examples{
req_auth_github <- function(req) {
  req_oauth_device(
    req,
    client = example_github_client(),
    auth_url = "https://github.com/login/device/code"
  )
}

request("https://api.github.com/user") \%>\%
  req_auth_github()
}
