% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-jwt.R
\name{req_oauth_jwt}
\alias{req_oauth_jwt}
\title{OAuth authentication with a JWT}
\usage{
req_oauth_jwt(
  req,
  app,
  claims,
  signature,
  signature_params = list(),
  scope = NULL,
  token_params = list()
)
}
\arguments{
\item{req}{A \link{request}.}

\item{app}{An \code{\link[=oauth_app]{oauth_app()}}.}

\item{signature}{Function use to sign \code{claim_set}, e.g. \code{\link[=jwt_sign_rs256]{jwt_sign_rs256()}}.}

\item{signature_params}{Additional arguments passed to \code{signature}}

\item{scope}{Scopes to be requested from the resource owner.}

\item{token_params}{List containing additional parameters passed to \code{oauth_flow_access_token()}.}
}
\description{
This uses \code{\link[=oauth_flow_jwt]{oauth_flow_jwt()}} to generate an access token which is then
used to authenticate the request with \code{\link[=req_auth_bearer_token]{req_auth_bearer_token()}}.
The token is cached in memory.
}
