% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-url-path.R
\name{req_url_path}
\alias{req_url_path}
\alias{req_url_path_append}
\title{Modify URL path}
\usage{
req_url_path(req, ...)

req_url_path_append(req, ...)
}
\arguments{
\item{req}{A httr2 \link{request} object.}

\item{...}{For \code{req_url_path()} and \code{req_url_path_append()}: A sequence of
path components that will be combined with \code{/}.}
}
\value{
A modified HTTP \link{request}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{req_url_path()} modifies the path.
\item \code{req_url_path_append()} adds to the path.
}

Please use \code{\link[=req_url_relative]{req_url_relative()}} instead.
}
\examples{
req <- request("http://example.com/a/b/c")

# OLD
req |> req_url_path("/d/e/f")
req |> req_url_path_append("d/e/f")

# NEW
req |> req_url_relative("/d/e/f")
req |> req_url_relative("d/e/f")
}
