% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{request}
\alias{request}
\title{Create a new HTTP request}
\usage{
request(base_url)
}
\arguments{
\item{base_url}{Base URL for request.}
}
\description{
To perform a HTTP request, first create a request object with \code{request()},
then define its behaviour with \code{req_} functions, then perform the request
and fetch the response with \code{\link[=req_perform]{req_perform()}}.
}
\examples{
request("http://r-project.org")
}
