% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-headers.R
\name{resp_date}
\alias{resp_date}
\title{Extract request date from response}
\usage{
resp_date(resp)
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[=req_perform]{req_perform()}}.}
}
\description{
All responses contain a request date in the \code{Date} header; if not provided
by the server will be automatically added by httr2.
}
