% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-headers.R
\name{resp_link_url}
\alias{resp_link_url}
\title{Parse link url from a response}
\usage{
resp_link_url(resp, rel)
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[=req_perform]{req_perform()}}.}

\item{rel}{"rel" value for which to retrieve url}
}
\description{
This parses the \code{Link} header, extracting the url corresponding to the
specified \code{rel}. Returns \code{NULL} if not present.
}
\examples{
resp <- request("https://api.github.com/search/code") \%>\%
  req_url_query(q = "addClass user:mozilla") \%>\%
  req_perform()
resp_link_url(resp, "next")
}
