% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{resp_raw}
\alias{resp_raw}
\title{Show the raw response}
\usage{
resp_raw(resp)
}
\arguments{
\item{resp}{An HTTP \link{response}}
}
\value{
\code{resp} (invisibly).
}
\description{
This function reconstructs the HTTP message that httr2 received from the
server. It's unlikely to be exactly byte-for-byte identical (because most
servers compress at least the body, and HTTP/2 can also compress the
headers), but it conveys the same information.
}
\examples{
resp <- request(test_server()) \%>\%
  req_url_path("/json") \%>\%
  req_perform()
resp \%>\% resp_raw()
}
