% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-request.R
\name{resp_request}
\alias{resp_request}
\title{Find the request responsible for a response}
\usage{
resp_request(resp)
}
\arguments{
\item{resp}{A httr2 \link{response} object, created by \code{\link[=req_perform]{req_perform()}}.}
}
\description{
To make debugging easier, httr2 includes the request that was used to
generate every response. You can use this function to access it.
}
\examples{
req <- request(example_url())
resp <- req_perform(req)
resp_request(resp)
}
