% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-timing.R
\name{resp_timing}
\alias{resp_timing}
\title{Extract information about the timing of the response}
\usage{
resp_timing(resp)
}
\arguments{
\item{resp}{A httr2 \link{response} object, created by \code{\link[=req_perform]{req_perform()}}.}
}
\value{
named numeric vector of timing information
}
\description{
The underlying curl library measures how long different components of the
request take to complete. This function retrieves that information.
}
\details{
The names of the elements in this vector correspond to the names used
in libcurl's \code{curl_easy_getinfo()} API. For example, \code{"namelookup"} is the
time returned by requesting \code{CURLINFO_NAMELOOKUP_TIME}. Refer to \href{https://curl.se/libcurl/c/curl_easy_getinfo.html}{curl documentation} for explanations of what these measure. The most useful
component is likely \code{"total"}, the overall time in seconds to complete the
request including any redirects followed.
}
\examples{
req <- request(example_url())
resp <- req_perform(req)
resp_timing(resp)
}
