% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-url.R
\name{resp_url}
\alias{resp_url}
\alias{resp_url_path}
\alias{resp_url_query}
\title{Get URL/components from the response}
\usage{
resp_url(resp)

resp_url_path(resp)

resp_url_query(resp)
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[=req_perform]{req_perform()}}.}
}
\description{
\itemize{
\item \code{resp_url()} returns the complete url.
\item \code{resp_url_path()} returns the path component.
\item \code{resp_url_query()} returns the query component as a named list.
}
}
\examples{
req <- request("https://httr2.r-lib.org?hello=world")

resp <- req_perform(req)
resp \%>\% resp_url()
resp \%>\% resp_url_path()
resp \%>\% resp_url_query()
}
