% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate-responses.R
\name{resps_combine}
\alias{resps_combine}
\alias{resps_responses}
\alias{resps_errors}
\title{Tools for working with lists of responses}
\usage{
resps_combine(resps, resp_data)

resps_responses(resps)

resps_errors(resps)
}
\arguments{
\item{resps}{A list of responses (possibly including errors).}

\item{resp_data}{A function with one argument \code{resp} that parses the
response and returns a list with the field \code{data} and other fields needed
to create the request for the next page.
\code{req_perform_iteratively()} combines all \code{data} fields via \code{\link[vctrs:vec_c]{vctrs::vec_c()}}
and returns the result.
Other fields that might be needed are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* `next_url` for `paginate_next_url()`.
* `next_token` for `paginate_next_token()`.
}\if{html}{\out{</div>}}}
}
\description{
\itemize{
\item \code{resps_combine()} combines the data from each responses into a single
object.
\item \code{resps_response()} returns all successful responses.
\item \code{resps_error()} returns all errors.
}
}
