% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate-helpers.R
\name{signal_total_pages}
\alias{signal_total_pages}
\title{Signal total number pages}
\usage{
signal_total_pages(n)
}
\arguments{
\item{n}{Total number of pages.}
}
\description{
To be called within a \code{next_req} callback function used with
\code{\link[=req_perform_iteratively]{req_perform_iteratively()}}
}
\keyword{internal}
