% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-throttle.R
\name{throttle_status}
\alias{throttle_status}
\title{Display internal throttle status}
\usage{
throttle_status()
}
\value{
A data frame with one row per token bucket and four columns:
\itemize{
\item The \code{realm}.
\item The \code{capacity} of the bucket.
\item Number of \code{tokens} remaining in the bucket.
\item Time \code{to_wait} in seconds for next token.
}
}
\description{
Sometimes useful for debugging.
}
\keyword{internal}
