% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url_build}
\alias{url_build}
\title{Build a string from a URL object}
\usage{
url_build(url)
}
\arguments{
\item{url}{An URL object created by \link{url_parse}.}
}
\description{
This is the inverse of \code{\link[=url_parse]{url_parse()}}, taking a parsed URL object and
turning it back into a string.
}
\seealso{
Other URL manipulation: 
\code{\link{url_modify}()},
\code{\link{url_parse}()}
}
\concept{URL manipulation}
