% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url_parse}
\alias{url_parse}
\title{Parse a URL}
\usage{
url_parse(url)
}
\arguments{
\item{url}{A string to parse.}
}
\value{
A URL, i.e. a S3 object with class \code{httr2_url} and elements
\code{scheme}, \code{hostname}, \code{username}, \code{password}, \code{port}, \code{path}, \code{query}, and
\code{fragment}.
}
\description{
\code{url_parse()} parses a URL into its component pieces, powered by
\code{\link[curl:curl_parse_url]{curl::curl_parse_url()}}. See \href{https://datatracker.ietf.org/doc/html/rfc3986}{RFC 3986} for the details of the
parsing algorithm.
}
\examples{
url_parse("http://google.com/")
url_parse("http://google.com:80/")
url_parse("http://google.com:80/?a=1&b=2")
url_parse("http://username@google.com:80/path;test?a=1&b=2#40")
}
\seealso{
Other URL manipulation: 
\code{\link{url_build}()},
\code{\link{url_modify}()}
}
\concept{URL manipulation}
