% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-mock.R
\name{with_mock}
\alias{with_mock}
\alias{local_mock}
\title{Temporarily mock requests}
\usage{
with_mock(mock, code)

local_mock(mock, env = caller_env())
}
\arguments{
\item{mock}{A single argument function called with a request object.
It should return either \code{NULL} (if it doesn't want to handle the request)
or a \link{response} (if it does).}

\item{code}{Code to execute in the temporary environment.}

\item{env}{Environment to use for scoping changes.}
}
\description{
Mocking allows you to selectively and temporarily replace the response
you would typically receive from a request with your own code. It's
primarily used for testing.
}
