% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content-type.R
\name{check_resp_content_type}
\alias{check_resp_content_type}
\title{Check the content type of a response}
\usage{
check_resp_content_type(resp, types, call = caller_env())
}
\arguments{
\item{resp}{A response object.}

\item{types}{A character vector of valid content types.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Returns \code{NULL} invisibly, or errors.
}
\description{
A different content type than expected often leads to an error in parsing
the response body. This function checks that the content type of the response
is as expected and fails otherwise.
}
\examples{
resp <- response(headers = list(`content-type` = "application/json"))
check_resp_content_type(resp, "application/json")
try(check_resp_content_type(resp, "application/xml"))

# `types` can also specify multiple valid types
check_resp_content_type(resp, c("application/xml", "application/json"))
}
