% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl.R
\name{curl_translate}
\alias{curl_translate}
\alias{curl_help}
\title{Translate curl syntax to httr2}
\usage{
curl_translate(cmd)

curl_help()
}
\arguments{
\item{cmd}{Call to curl}
}
\description{
The curl command line tool is commonly to demonstrate HTTP APIs and can
easily be be generated from browser developer tools. \code{curl_translate()}
implements a partial, but frequently used, subset of curl functionality
allowing you to automatically translate curl invocations into their httr2
equivalents.

Use \code{curl_help()} to see the supported options, and \code{curl_translate()}
to translate a curl invocation copy and pasted from elsewhere.
}
\examples{
curl_translate("curl http://example.com")
curl_translate("curl http://example.com -X DELETE")
curl_translate("curl http://example.com --header A:1 --header B:2")
curl_translate("curl http://example.com --verbose")
}
