% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{example_url}
\alias{example_url}
\title{URL to a local server that's useful for tests and examples}
\usage{
example_url()
}
\description{
Requires the webfakes package to be installed. It has the following endpoints:
}
\details{
\itemize{
\item all the ones from the \code{\link[webfakes:httpbin_app]{webfakes::httpbin_app()}}
\item \verb{/iris}: paginate through the iris dataset. It has the query parameters
\code{page} and \code{limit} to control the pagination.
}
}
\keyword{internal}
