% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_client_logout}
\alias{handle_oauth_client_logout}
\title{Handle OAuth Client Logout}
\usage{
handle_oauth_client_logout(req, client_config)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client_config}{A list of client configurations used for OAuth.}
}
\value{
An HTTP response object or \code{NULL} if no client is matched.
}
\description{
This function manages the logout process for OAuth clients. It deletes the
relevant cookies and redirects the user back to the app.
}
\keyword{internal}
