% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-shiny-http-handlers.R
\name{handle_oauth_client_logout_delete_cookies}
\alias{handle_oauth_client_logout_delete_cookies}
\title{Handle OAuth Client Logout - Delete Cookies}
\usage{
handle_oauth_client_logout_delete_cookies(req, client)
}
\arguments{
\item{req}{A \code{shiny} request object.}

\item{client}{A single client configuration object.}
}
\value{
An HTTP response object or \code{NULL} if the request path does not match
the client's logout path.
}
\description{
This function deletes the cookies associated with a specific OAuth client
during the logout process. It then redirects the user back to the app.
}
\keyword{internal}
