% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2-package.R
\docType{package}
\name{httr2-package}
\alias{httr2}
\alias{httr2-package}
\title{httr2: Perform HTTP Requests and Process the Responses}
\description{
httr2 is a modern re-imagining on httr designed around a
    pipeable interface and solving more of the problems that API wrappers
    need.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://httr2.r-lib.org}
  \item \url{https://github.com/r-lib/httr2}
  \item Report bugs at \url{https://github.com/r-lib/httr2/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
