% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2-package.R
\docType{package}
\name{httr2-package}
\alias{httr2}
\alias{httr2-package}
\title{httr2: Perform HTTP Requests and Process the Responses}
\description{
Tools for creating and modifying HTTP requests, then performing them and processing the results. 'httr2' is a modern re-imagining of 'httr' that uses a pipe-based interface and solves more of the problems that API wrapping packages face.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://httr2.r-lib.org}
  \item \url{https://github.com/r-lib/httr2}
  \item Report bugs at \url{https://github.com/r-lib/httr2/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
  \item Maximilian Girlich [contributor]
}

}
\keyword{internal}
