% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-fetch.R
\name{last_response}
\alias{last_response}
\alias{last_request}
\title{Retrieve most recent request/response}
\usage{
last_response()

last_request()
}
\description{
These functions retrieve the most recent request made by httr2 and
the response it received, to facilitate debugging problems \emph{after} they
occur. If the request did not succeed (or no requests have been made)
\code{last_response()} will be \code{NULL}.
}
