% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_req_perform.R, R/req-perform.R
\name{last_chunked_responses}
\alias{last_chunked_responses}
\alias{last_chunk}
\alias{last_response}
\alias{last_request}
\title{Retrieve most recent request/response}
\usage{
last_chunked_responses()

last_chunk()

last_response()

last_request()
}
\value{
An HTTP \link{response}/\link{request}.
}
\description{
These functions retrieve the most recent request made by httr2 and
the response it received, to facilitate debugging problems \emph{after} they
occur. If the request did not succeed (or no requests have been made)
\code{last_response()} will be \code{NULL}.
}
\examples{
invisible(request("http://httr2.r-lib.org") \%>\% req_perform())
last_request()
last_response()
}
