% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-app.R
\name{oauth_app}
\alias{oauth_app}
\alias{oauth_client}
\title{Create an OAuth app}
\usage{
oauth_app(
  client,
  endpoints,
  auth = c("body", "header", "jwt_rs256"),
  auth_params = list()
)

oauth_client(client_id, client_secret = NULL, name = NULL)
}
\arguments{
\item{client}{A OAuth client created with \code{oauth_client()}.}

\item{endpoints}{A named character vector of endpoints. The precise endpoints
required depend on the flow that you'll use, but all require a \code{token}
endpoint that returns an access token.}

\item{auth}{Authentication mechanism used by the client to prove itself to
the API. Can be one of three built-in methods ("body", "header", or "jwt"),
or a function that will be called with arguments \code{req}, \code{client}, and
the contents of \code{auth_params}.

The most common mechanism in the wild is \code{"body"} where the \code{client_id} and
(optionally) \code{client_secret} are added to the body. \code{"header"} sends in
\code{client_id} and \code{client_secret} in HTTP Authorization header. \code{"jwt_s256"}
generate a JWT claim set signed with RS256.

See \code{\link[=oauth_client_req_auth]{oauth_client_req_auth()}} for more details.}

\item{auth_params}{Additional parameters passed to the function specified
by \code{auth}.}

\item{client_id}{Client identifier.}

\item{client_secret}{Client secret. This is technically confidential so you
should avoid storing it in source code where possible. However, many APIs
require it in order to provide a user friendly authentication experience,
and the risks of including it are usually low. To make things a little
safer, I recommend using \code{\link[=obfuscate]{obfuscate()}} to store an obfuscated version of
the source.}

\item{name}{Optional name for the client. Used when generating cache
directory. If \code{NULL}, generated from hash of app hostname and
\code{client_id}.}
}
\description{
An OAuth app is the combination of a client, a set of endpoints
(i.e. urls where various requests should be sent), and an authentication
mechanism. A client consists of at least a \code{client_id}, and also often
a \code{client_secret}. You'll get these values when you create the client on
the API's website.
}
\examples{
# Set up an OAuth app for MS Azure that authenticates with a private key
# https://bit.ly/3wc1WJB

client_id <- "my_client_id"
claims <- list(
  aud = "https://login.microsoftonline.com/{tenant}/v2.0",
  iss = client_id,
  sub = client_id
)
oauth_app(
  oauth_client("id"),
  endpoints = c(
    token = "https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token"
  ),
  auth = "jwt_rs256",
  auth_params = list(
    claims = claims,
    key = "path/to/private/key"
  )
)
}
