% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth-flow-client-credentials.R
\name{oauth_flow_client_credentials}
\alias{oauth_flow_client_credentials}
\title{OAuth flow: client credentials}
\usage{
oauth_flow_client_credentials(app, scope = NULL, token_params = list())
}
\arguments{
\item{app}{An \code{\link[=oauth_app]{oauth_app()}}.}

\item{scope}{Scopes to be requested from the resource owner.}

\item{token_params}{List containing additional parameters passed to \code{oauth_flow_access_token()}.}
}
\description{
This function implements the OAuth client credentials flow, as defined
by \href{https://datatracker.ietf.org/doc/html/rfc6749#section-4.4}{rfc6749},
Section 4.4. It is used to allow the client to access resources that it
controls directly, not on behalf of an user.
}
\seealso{
Other OAuth flows: 
\code{\link{oauth_flow_auth_code}()},
\code{\link{oauth_flow_device}()},
\code{\link{oauth_flow_jwt}()},
\code{\link{oauth_flow_password}()},
\code{\link{oauth_flow_refresh}()}
}
\concept{OAuth flows}
