% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{obfuscate}
\alias{obfuscate}
\alias{obfuscated}
\title{Obfuscate mildly secret information}
\usage{
obfuscate(x)

obfuscated(x)
}
\arguments{
\item{x}{A string to obfuscate, or mark as obfuscated.}
}
\value{
\code{obfuscate()} prints the \code{obfuscated()} call to include in your
code. \code{obfuscated()} returns an S3 class marking the string as obfuscated
so it can be unobfuscated when needed.
}
\description{
This pair of functions provides a way to obfuscate mildly confidential
information, like OAuth client secrets. The secret can not be revealed
from your source code, but a good R programmer could still figure it
out with a little effort. The main goal is to protect against scraping;
there's no way for an automated tool to grab your obfuscated secrets.

Because un-obfuscation happens at the last possible instant, \code{obfuscated()}
only works in limited locations:
\itemize{
\item The \code{secret} argument to \code{\link[=oauth_client]{oauth_client()}}
\item Elements of the \code{data} argument to \code{\link[=req_body_form]{req_body_form()}}, \code{req_body_json()},
and \code{req_body_multipart()}.
}
}
\examples{
obfuscate("good morning")
obfuscated("dV-4_vKoUp90pP_M")
}
