% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate_req_perform}
\alias{paginate_req_perform}
\alias{paginate_next_request}
\title{Perform a paginated request}
\usage{
paginate_req_perform(req, max_pages = 20L, progress = TRUE)

paginate_next_request(resp, req)
}
\arguments{
\item{req}{A \link{request}.}

\item{max_pages}{The maximum number of pages to request.}

\item{progress}{Display a progress bar?}

\item{resp}{An HTTP \link{response}.}
}
\value{
A list of responses.
}
\description{
Perform a paginated request
}
\examples{
req_pokemon <- request("https://pokeapi.co/api/v2/pokemon") \%>\%
  req_paginate_next_url(
    next_url = function(resp) resp_body_json(resp)[["next"]],
    page_size = 150,
    set_page_size = function(req, page_size) {
      req_url_query(req, limit = page_size)
    },
    n_pages = function(resp) {
      total <- resp_body_json(resp)$count
      ceiling(total / page_size)
    }
  )

responses <- paginate_req_perform(req_pokemon)
}
