% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{req}
\alias{req}
\title{Create a new HTTP request}
\usage{
req(base_url)
}
\arguments{
\item{base_url}{Base URL for request.}
}
\description{
Create a new HTTP request
}
\examples{
req("http://r-project.org")
}
