% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-auth.R
\name{req_auth_basic}
\alias{req_auth_basic}
\title{Authenticate request with HTTP basic authentication}
\usage{
req_auth_basic(req, username, password = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{username}{User name.}

\item{password}{Password. You avoid entering the password directly when
calling this function as it will be captured by \code{.Rhistory}. Instead,
leave it unset and the default behaviour will prompt you for it
interactively.}
}
\description{
This sets the Authorization header. See details at
\url{https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Authorization}.
}
\examples{
req <- request("http://example.com") \%>\% req_auth_basic("hadley", "SECRET")
req
req \%>\% req_dry_run()

# httr2 does its best to redact the Authorization header so that you don't
# accidentally confidential data. If you need to see exactly what's happening
# you can use redact = FALSE
print(req, redact = FALSE)
req \%>\% req_dry_run(redact = FALSE)

# Note that the authorization header is not encrypted and the
# password can easily be restored:
rawToChar(jsonlite::base64_dec("aGFkbGV5IDogU0VDUkVU"))
# This means that you should be careful when sharing any code that
# reveals the Authorization header
}
