% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-auth.R
\name{req_auth_bearer_token}
\alias{req_auth_bearer_token}
\title{Authenticate request with bearer token}
\usage{
req_auth_bearer_token(req, token)
}
\arguments{
\item{req}{A \link{req}uest.}

\item{token}{A bearer token}
}
\description{
A bearer token gives the bearer access to confidential resources
(so you should keep them secure like you would with a user name and
password). They are usually produced by some large authentication scheme
(like the various OAuth 2.0 flows), but you are sometimes given then
directly.
}
\seealso{
\href{https://datatracker.ietf.org/doc/html/rfc6750}{RFC750}
The OAuth 2.0 Authorization Framework: Bearer Token Usage
}
