% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-fetch.R
\name{req_dry_run}
\alias{req_dry_run}
\title{Perform a dry run}
\usage{
req_dry_run(req, quiet = !is_interactive())
}
\arguments{
\item{req}{A \link{req}uest.}

\item{quiet}{If \code{TRUE} doesn't print anything.}
}
\value{
Invisibly, a list containing information about the request,
including \code{method}, \code{path}, and \code{headers}.
}
\description{
This shows you exactly what httr2 will send to the server, without
actually sending anything. It requires the httpuv package because it
works by sending the real HTTP request to a local webserver, thanks to
the magic of \code{\link[curl:curl_echo]{curl::curl_echo()}}.
}
\examples{
req("http://google.com") \%>\% req_dry_run()
}
