% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-fetch.R
\name{req_fetch}
\alias{req_fetch}
\title{Perform a request, fetching data back to R}
\usage{
req_fetch(req, path = NULL, handle = NULL)
}
\arguments{
\item{req}{A \link{req}uest.}

\item{path}{Optionally, path to save body of request. This is useful for
large responses since it avoids storing the response in memory.}

\item{handle}{Advanced use only; a curl handle.}
}
\value{
Returns an HTTP response.
}
\description{
After preparing a request, call \code{req_fetch()} to perform it, fetching
the results back to R.
}
\examples{
req("https://google.com") \%>\%
  req_fetch()
}
