% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-fetch.R
\name{req_fetch}
\alias{req_fetch}
\title{Simulate a request}
\usage{
req_fetch(req, path = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{path}{Optionally, path to save body of request. This is useful for
large responses since it avoids storing the response in memory.}
}
\value{
Returns an HTTP response with successful status code. Will
throw an error for 4xx and 5xx responses; override with \code{\link[=req_error]{req_error()}}.
}
\description{
After preparing a request, call \code{req_fetch()} to perform it, fetching
the results back to R. One call to \code{req_fetch()} may perform multiple
HTTP requests:
\itemize{
\item If the \code{url} is redirected with a 301, 302, 303, or 307, curl will
automatically follow the \code{Location} header to the new location.
\item If you have configured retries with \code{\link[=req_retry]{req_retry()}} and the request
fails with a transient problem, \code{req_fetch()} will try again after
waiting a bit. See \code{\link[=req_retry]{req_retry()}} for details.
}
}
\examples{
request("https://google.com") \%>\%
  req_fetch()
}
