% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-method.R
\name{req_get_method}
\alias{req_get_method}
\title{Retrieve the method that a request will use}
\usage{
req_get_method(req)
}
\arguments{
\item{req}{A httr2 \link{request} object.}
}
\description{
Mimics the algorithm that curl uses to determine the method.
}
\examples{
req <- request(example_url())
req_get_method(req)
req_get_method(req |> req_body_raw("abc"))
req_get_method(req |> req_method("DELETE"))
req_get_method(req |> req_method("HEAD"))
}
\keyword{internal}
