% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-headers.R
\name{req_headers}
\alias{req_headers}
\title{Modify request headers}
\usage{
req_headers(.req, ...)
}
\arguments{
\item{.req}{A \link{request}.}

\item{...}{Name-value pairs of headers and their values. Note that setting a
header to \code{NULL} will not necessarily remove it, but will reset it
to httr2's default value. To remove a header, set it to \code{""}.}
}
\description{
\code{req_headers()} allows you to set the value of any header.
}
\examples{
req <- request("http://example.com")
# Setting Accept to NULL uses curl's default:
req \%>\%
  req_headers(Accept = NULL) \%>\%
  req_dry_run()

# Setting it to "" removes it:
req \%>\%
  req_headers(Accept = "") \%>\%
  req_dry_run()
}
