% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-method.R
\name{req_method}
\alias{req_method}
\title{Modify request method}
\usage{
req_method(req, method)
}
\arguments{
\item{req}{A \link{request}.}

\item{method}{Custom HTTP method}
}
\description{
Use this function to use a custom HTTP method like "HEAD",
"DELETE", "PATCH", "UPDATE", or "OPTIONS".
}
\examples{
request("http://httpbin.org") \%>\% req_method("PATCH")
request("http://httpbin.org") \%>\% req_method("PUT")
request("http://httpbin.org") \%>\% req_method("HEAD")
}
