% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-options.R
\name{req_options}
\alias{req_options}
\title{Set arbitrary curl options in request}
\usage{
req_options(.req, ...)
}
\arguments{
\item{.req}{A \link{req}uest.}

\item{...}{Name-value pairs. The name should be a valid curl option,
as found in \code{\link[curl:curl_options]{curl::curl_options()}}.}
}
\description{
\code{req_options()} is for expert use only; it allows you to directly set
libcurl options to access features that are otherwise not available in
httr2.
}
\keyword{internal}
