% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{req_paginate}
\alias{req_paginate}
\alias{req_paginate_next_url}
\alias{req_paginate_offset}
\alias{req_paginate_next_token}
\title{Pagination}
\usage{
req_paginate(req, next_request, n_pages = NULL)

req_paginate_next_url(req, next_url, n_pages = NULL)

req_paginate_offset(req, offset, page_size, n_pages = NULL)

req_paginate_next_token(req, set_token, next_token, n_pages = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{next_request}{A callback function that takes a two arguments (the
original request and the response) and returns:
\itemize{
\item a new \link{request} to request the next page or
\item \code{NULL} if there is no next page.
}}

\item{n_pages}{A function that extracts the next token from the \link{response}.}

\item{next_url}{A function that extracts the next url from the \link{response}.}

\item{offset}{A function that applies that applies the new offset to the
request. It takes two arguments: a \link{request} and an integer offset.}

\item{page_size}{A parameter object that specifies how the page size is added
to the request.}

\item{set_token}{A function that applies that applies the new token to the
request. It takes two arguments: a \link{request} and the new token.}

\item{next_token}{A function that extracts the next token from the \link{response}.}
}
\value{
A modified HTTP \link{request}.
}
\description{
Pagination
}
\examples{
request("https://pokeapi.co/api/v2/pokemon") \%>\%
  req_url_query(limit = 150) \%>\%
  req_paginate_next_url(
    next_url = function(resp) resp_body_json(resp)[["next"]],
    n_pages = function(resp) {
      calculate_n_pages(
        page_size = 150,
        total = resp_body_json(resp)$count
      )
    }
  )
}
