% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{req_paginate}
\alias{req_paginate}
\alias{req_paginate_next_url}
\alias{req_paginate_offset}
\alias{req_paginate_next_token}
\title{Pagination}
\usage{
req_paginate(req, next_request, n_pages = NULL)

req_paginate_next_url(req, next_url, n_pages = NULL)

req_paginate_offset(req, offset, page_size, n_pages = NULL)

req_paginate_next_token(req, set_token, next_token, n_pages = NULL)
}
\arguments{
\item{req}{A \link{request}.}

\item{next_request}{A callback function that takes a two arguments (the
original request and the response) and returns:
\itemize{
\item a new \link{request} to request the next page or
\item \code{NULL} if there is no next page.
}}

\item{n_pages}{A function that extracts the total number of pages from
the \link{response}.}

\item{next_url}{A function that extracts the url to the next page from the
\link{response}.}

\item{offset}{A function that applies the new offset to the request. It takes
two arguments: a \link{request} and an integer offset.}

\item{page_size}{A whole number that specifies the page size i.e. the number
of elements per page.}

\item{set_token}{A function that applies the new token to the request. It
takes two arguments: a \link{request} and the new token.}

\item{next_token}{A function that extracts the next token from the \link{response}.}
}
\value{
A modified HTTP \link{request}.
}
\description{
Use \code{req_paginate()} to specify how to request the next page in a paginated
API. There are also helpers for common pagination patterns:
}
\details{
\itemize{
\item \code{req_paginate_next_url()} when the response contains a link to the next
page.
\item \code{req_paginate_offset()} when the request describes the offset i.e.
at which element to start and the page size.
\item \code{req_paginate_next_token()} when the response contains a token
that is used to describe the next page.
}

Use \code{\link[=paginate_req_perform]{paginate_req_perform()}} to fetch all pages.
If you need more control use a combination of \code{\link[=req_perform]{req_perform()}} and
\code{\link[=paginate_next_request]{paginate_next_request()}} to iterate through the pages yourself.
}
\examples{
page_size <- 150

request("https://pokeapi.co/api/v2/pokemon") \%>\%
  req_url_query(limit = page_size) \%>\%
  req_paginate_next_url(
    next_url = function(resp) resp_body_json(resp)[["next"]],
    n_pages = function(resp) {
      total <- resp_body_json(resp)$count
      ceiling(total / page_size)
    }
  )
}
\seealso{
\code{\link[=paginate_req_perform]{paginate_req_perform()}} to fetch all pages. \code{\link[=paginate_next_request]{paginate_next_request()}}
to generate the request to the next page.
}
