% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{req_paginate}
\alias{req_paginate}
\alias{req_paginate_next_url}
\alias{req_paginate_offset}
\alias{req_paginate_next_token}
\title{Pagination}
\usage{
req_paginate(req, next_request, page_size = NULL, total = NULL)

req_paginate_next_url(req, next_url, ..., page_size = NULL, total = NULL)

req_paginate_offset(req, offset, page_size, total = NULL)

req_paginate_next_token(
  req,
  token_field,
  next_token_field,
  page_size = NULL,
  total = NULL
)
}
\arguments{
\item{req}{A \link{request}.}

\item{next_request}{A callback function that takes a two arguments (the
original request and the response) and returns:
\itemize{
\item a new \link{request} to request the next page or
\item \code{NULL} if there is no next page.
}}

\item{page_size}{A parameter object that specifies how the page size is added
to the request.}

\item{total}{A character that specifies the path where in the body the field
with the total number of elements is stored.}

\item{next_url}{A character that specifies the path where in the body the field
with the next url of the next page is stored.}

\item{offset}{A parameter object that specifies how the offset is added to
the request.}

\item{token_field}{A parameter object that specifies how the next token is
added to the request.}

\item{next_token_field}{A character that specifies the path where in the body
the field with the token of the next page is stored.}
}
\value{
A modified HTTP \link{request}.
}
\description{
Pagination
}
\examples{
request("https://pokeapi.co/api/v2/pokemon") \%>\%
  req_paginate_next_url(
    "next",
    page_size = in_query("limit", 150L),
    total = "count"
  )
}
