% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-perform-stream.R
\name{req_perform_connection}
\alias{req_perform_connection}
\alias{resp_stream_sse}
\title{Perform a request and return a streaming connection}
\usage{
req_perform_connection(req, blocking = TRUE)

resp_stream_sse(resp)
}
\arguments{
\item{req}{A \link{request}.}

\item{blocking}{When retrieving data, should the connection block and wait
for the desired information or immediately return what it has?}

\item{resp}{A response object.}
}
\description{
Use \code{req_perform_connection()} to perform a request that includes a
connection as the body of the response, then use \code{resp_steam_sse()} and
friends to retrieve data a chunk at a time..

This is an alternative interface to \code{\link[=req_perform_stream]{req_perform_stream()}} that returns a
connection that you can pull from data, rather than callbacks that are called
as the data streams in. This is useful if you want to do other work in
between streaming inputs.
}
