% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-options.R
\name{req_proxy}
\alias{req_proxy}
\title{Use a proxy for a request}
\usage{
req_proxy(
  req,
  url,
  port = NULL,
  username = NULL,
  password = NULL,
  auth = "basic"
)
}
\arguments{
\item{req}{A \link{request}.}

\item{url, port}{Location of proxy.}

\item{username, password}{Login details for proxy, if needed.}

\item{auth}{Type of HTTP authentication to use. Should be one of the
following: basic, digest, digest_ie, gssnegotiate, ntlm, any.}
}
\description{
Use a proxy for a request
}
\examples{
# See http://www.hidemyass.com/proxy-list for a list of public proxies
# to test with
# request("http://had.co.nz") \%>\%
#   req_proxy("64.251.21.73", 8080) \%>\%
#   req_perform()
}
