% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-proxy.R
\name{req_proxy}
\alias{req_proxy}
\title{Use a proxy to connect to the internet.}
\usage{
req_proxy(
  req,
  url,
  port = NULL,
  username = NULL,
  password = NULL,
  auth = "basic"
)
}
\arguments{
\item{url, port}{location of proxy}

\item{username, password}{login details for proxy, if needed}

\item{auth}{type of HTTP authentication to use. Should be one of the
following: basic, digest, digest_ie, gssnegotiate, ntlm, any.}
}
\description{
Use a proxy to connect to the internet.
}
\examples{
# See http://www.hidemyass.com/proxy-list for a list of public proxies
# to test with
# request("http://had.co.nz") \%>\%
#   req_proxy("64.251.21.73", 8080) \%>\%
#   req_perform()
}
