% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-throttle.R
\name{req_throttle}
\alias{req_throttle}
\title{Throttle a request by automatically adding a delay}
\usage{
req_throttle(req, requests_per_second)
}
\arguments{
\item{req}{A \link{req}uest.}

\item{requests_per_second}{Maximum number of requests per second - usually
easier to specify as a fraction, \code{number_of_requests / number_of_seconds}.
For example, if you want to make 15 requests per minute, write \code{15 / 60}.}
}
\description{
Throttling a request ensures that it never makes more than the specified
number of requests per second.
}
