% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-url.R
\name{req_url}
\alias{req_url}
\alias{req_url_query}
\alias{req_url_path}
\alias{req_url_path_append}
\title{Modify request URL}
\usage{
req_url(req, url)

req_url_query(req, ...)

req_url_path(req, path)

req_url_path_append(req, path)
}
\arguments{
\item{req}{A \link{request}.}

\item{url}{New URL; completely replaces existing.}

\item{...}{Name-value pairs that provide query parameters.}

\item{path}{Path to replace or append to existing path.}
}
\value{
A modified HTTP \link{request}.
}
\description{
\itemize{
\item \code{req_url()} replaces the entire url
\item \code{req_url_query()} modifies the components of the query
\item \code{req_url_path()} modifies the path
\item \code{req_url_path_append()} adds to the path
}
}
\examples{
req <- request("http://example.com")

# Change url components
req \%>\%
  req_url_path_append("a") \%>\%
  req_url_path_append("b") \%>\%
  req_url_path_append("search.html") \%>\%
  req_url_query(q = "the cool ice")

# Change complete url
req \%>\%
  req_url("http://google.com")
}
