% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-options.R
\name{req_user_agent}
\alias{req_user_agent}
\title{Modify request user agent}
\usage{
req_user_agent(req, ua)
}
\arguments{
\item{req}{A \link{req}uest.}

\item{ua}{A new user-agent string.}
}
\description{
This overrides the default user agent set by httr2 which includes the
version numbers of httr2, the curl package, and libcurl.
}
\examples{
req("http://example.com") \%>\% req_dry_run()
req("http://example.com") \%>\% req_user_agent("MyPackage") \%>\% req_dry_run()
}
