% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-options.R
\name{req_verbose}
\alias{req_verbose}
\title{Show verbose output when request is performed}
\usage{
req_verbose(
  req,
  header_out = TRUE,
  header_in = TRUE,
  data_out = TRUE,
  data_in = FALSE,
  info = FALSE,
  ssl = FALSE
)
}
\arguments{
\item{req}{A \link{req}uest.}

\item{header_out, header_in}{Show headers sent to/received from the server?}

\item{data_out, data_in}{Show data sent to/received from the server?}

\item{info}{Show informational text from curl. This is mainly useful
for debugging https and auth problems, so is disabled by default.}

\item{ssl}{Show data even when using a secure connection?}
}
\description{
\code{req_verbose()} uses the following prefixes to distinguish between
different components of the http messages:
\itemize{
\item \code{*} informative curl messages
\item \verb{->} headers sent (out)
\item \verb{>>} data sent (out)
\item \verb{*>} ssl data sent (out)
\item \verb{<-} headers received (in)
\item \verb{<<} data received (in)
\item \verb{<*} ssl data received (in)
}
}
