% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-headers.R
\name{resp_headers}
\alias{resp_headers}
\alias{resp_header}
\alias{resp_header_exists}
\title{Extract headers from a response}
\usage{
resp_headers(resp)

resp_header(resp, header)

resp_header_exists(resp, header)
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[=req_perform]{req_perform()}}.}

\item{header}{Header name (case insensitive)}
}
\description{
\itemize{
\item \code{resp_headers()} retrieves a list of all headers.
\item \code{resp_header()} retrieves a single header.
\item \code{resp_header_exists()} checks if a header is present.
}
}
