% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp.R
\name{resp_raw}
\alias{resp_raw}
\title{Show the raw response}
\usage{
resp_raw(resp)
}
\arguments{
\item{resp}{A HTTP \link{response}}
}
\description{
The reconstruct the HTTP message that httr2 received from the server.
It's unlikely to be exactly the same (because most servers compress at
least the body, and HTTP/2 can also compress the headers), but it conveys
the same information.
}
\examples{
resp <- request("https://httpbin.org/json") \%>\% req_perform()
resp \%>\% resp_raw()
}
