% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp-status.R
\name{resp_status}
\alias{resp_status}
\alias{resp_status_desc}
\alias{resp_is_error}
\title{Retrieve response status}
\usage{
resp_status(resp)

resp_status_desc(resp)

resp_is_error(resp)
}
\arguments{
\item{resp}{An HTTP response object, as created by \code{\link[=req_fetch]{req_fetch()}}.}
}
\description{
\itemize{
\item \code{http_status()} retrieves the numeric HTTP status code
\item \code{http_status_desc()} retrieves the brief textual description.
\item \code{http_is_error()} returns \code{TRUE} if the status code represents an error
(i.e. a 4xx or 5xx status).
}

In most cases you will not see 1xx or 3xx status codes, as these are
handled automatically by libcurl.
}
