% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-perform-stream.R
\name{resp_stream_raw}
\alias{resp_stream_raw}
\alias{resp_stream_lines}
\alias{resp_stream_sse}
\alias{close.httr2_response}
\title{Read a streaming body a chunk at a time}
\usage{
resp_stream_raw(resp, kb = 32)

resp_stream_lines(resp, lines = 1)

resp_stream_sse(resp)

\method{close}{httr2_response}(con, ...)
}
\arguments{
\item{resp, con}{A streaming \link{response} created by \code{\link[=req_perform_connection]{req_perform_connection()}}.}

\item{kb}{How many kilobytes (1024 bytes) of data to read.}

\item{lines}{How many lines to read}

\item{...}{Not used; included for compatibility with generic.}
}
\value{
\itemize{
\item \code{resp_stream_raw()}: a raw vector.
\item \code{resp_stream_lines()}: a character vector.
\item \code{resp_stream_sse()}: A list with components \code{type}, \code{data}, and \code{id}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\itemize{
\item \code{resp_stream_raw()} retrieves one or more bytes.
\item \code{resp_stream_lines()} retrieves one or more lines.
\item \code{resp_stream_sse()} retrieves a single
\href{https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events}{server-sent events}
from the stream. It current  only works with text mode connections so when
calling\code{req_perform_connection()} you must use \code{mode = "text"}.
}

\code{resp_stream_sse()} helps work with APIs that uses the
}
