% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate-responses.R
\name{resps_combine}
\alias{resps_combine}
\alias{resps_responses}
\alias{resps_errors}
\title{Tools for working with lists of responses}
\usage{
resps_combine(resps, resp_data)

resps_responses(resps)

resps_errors(resps)
}
\arguments{
\item{resps}{A list of responses (possibly including errors).}

\item{resp_data}{A function that takes a response (\code{resp}) and
returns the data foind inside that response as a vector or data frame.}
}
\description{
\itemize{
\item \code{resps_combine()} combines the data from each response into a single
vector.
\item \code{resps_response()} returns all successful responses.
\item \code{resps_error()} returns all errors.
}
}
