% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req-throttle.R
\name{throttle_status}
\alias{throttle_status}
\title{Display internal throttle status}
\usage{
throttle_status()
}
\value{
A data frame with two columns: the \code{realm} and time the
\code{last_request} was made.
}
\description{
Sometimes useful for debugging.
}
\keyword{internal}
