% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{url_modify}
\alias{url_modify}
\title{Modify a url}
\usage{
url_modify(
  url,
  scheme = NULL,
  hostname = NULL,
  username = NULL,
  password = NULL,
  port = NULL,
  path = NULL,
  query = NULL,
  fragment = NULL
)
}
\arguments{
\item{url}{A string or \href{url_parse}{parsed URL}.}

\item{scheme}{The scheme, typically either \code{http} or \code{https}.}

\item{hostname}{The hostname, e.g. \code{www.google.com} or \code{posit.co}.}

\item{username, password}{Username and password to embed in the URL.
Not generally recommended but needed for some legacy applications.}

\item{port}{An integer port number.}

\item{path}{The path, e.g. \verb{/search}. Paths must start with \code{/}, so this
will be automatically added if ommitted.}

\item{query}{Either a query string or a named list of query components.}

\item{fragment}{The fragment, e.g. \verb{#section-1}.}
}
\value{
An object the same type as `url``.
}
\description{
Modify components of a URL. The default value of each argument, \code{NULL},
means leave the component as is. If you want to remove a component,
set it to \code{""}. Note that setting \code{scheme} or \code{hostname} to \code{""} will
create a relative url.
}
\examples{
url_modify("http://hadley.nz", path = "about")
url_modify("http://hadley.nz", scheme = "https")
url_modify("http://hadley.nz/abc", path = "/cde")
url_modify("http://hadley.nz/abc", path = "")
url_modify("http://hadley.nz?a=1", query = "b=2")
url_modify("http://hadley.nz?a=1", query = list(c = 3))
}
\seealso{
Other URL manipulation: 
\code{\link{url_build}()},
\code{\link{url_parse}()}
}
\concept{URL manipulation}
